Add files via upload
A teacher in a school wants to find and display the grade of a student based on his/her
percentage score. The criterion for grades is as given below:

Score (both inclusive) Grade

Between 80 and 100 A
Between 73 and 79 B
Between 65 and 72 C
Between 0 and 64 D
Any other value Z

Assume that a percentage score is a whole number. Write a python program for the above
requirement.