Add files via upload
This is a guessing game. A random number will generate and you have to guess it. If you guess it wrong it will help you to guess lower or upper as per your guess.
When you put the right number it will show you the number of attempts you took to guess the number